Address PR comments (part 1)
