+++ src/pages/version/Tasks.tsx
@@ -1,16 +1,8 @@
 import { useEffect } from "react";
 import { useQuery } from "@apollo/client";
-import styled from "@emotion/styled";
-import Button from "@leafygreen-ui/button";
 import { Skeleton } from "antd";
 import { useParams, useLocation } from "react-router-dom";
-import { useVersionAnalytics } from "analytics";
-import { PageSizeSelector } from "components/PageSizeSelector";
-import { Pagination } from "components/Pagination";
-import { ResultCountLabel } from "components/ResultCountLabel";
-import { TableControlOuterRow, TableControlInnerRow } from "components/styles";
 import { pollInterval } from "constants/index";
-import { size } from "constants/tokens";
 import { useToastContext } from "context/toast";
 import { PatchTasksQuery, PatchTasksQueryVariables } from "gql/generated/types";
 import { GET_PATCH_TASKS } from "gql/queries";
@@ -18,6 +10,7 @@ import { usePolling } from "hooks";
 import { useUpdateURLQueryParams } from "hooks/useUpdateURLQueryParams";
 import { PatchTasksQueryParams } from "types/task";
 import { queryString, url, array } from "utils";
+import { TabTableControl } from "./TabTableControl";
 import { PatchTasksTable } from "./tasks/PatchTasksTable";
 
 const { toArray } = array;
@@ -32,7 +25,6 @@ export const Tasks: React.VFC<Props> = ({ taskCount }) => {
   const { id: versionId } = useParams<{ id: string }>();
 
   const { search } = useLocation();
-  const { sendEvent } = useVersionAnalytics(versionId);
   const dispatchToast = useToastContext();
 
   const updateQueryParams = useUpdateURLQueryParams();
@@ -66,50 +58,15 @@ export const Tasks: React.VFC<Props> = ({ taskCount }) => {
   usePolling(startPolling, stopPolling);
   const { patchTasks } = data || {};
 
-  const onClearAll = () => {
-    sendEvent({ name: "Clear all filter" });
-    updateQueryParams({
-      statuses: undefined,
-      baseStatuses: undefined,
-      taskName: undefined,
-      variant: undefined,
-      page: undefined,
-      sorts: defaultSortMethod,
-    });
-  };
-
   return (
     <>
-      <TableControlOuterRow>
-        <FlexContainer>
-          <ResultCountLabel
-            dataCyNumerator="current-task-count"
-            dataCyDenominator="total-task-count"
-            label="tasks"
-            numerator={patchTasks?.count}
-            denominator={taskCount}
-          />
-          <PaddedButton // @ts-expect-error
-            onClick={onClearAll}
-            data-cy="clear-all-filters"
-          >
-            Clear All Filters
-          </PaddedButton>
-        </FlexContainer>
-        <TableControlInnerRow>
-          <Pagination
-            data-cy="tasks-table-pagination"
-            pageSize={limit}
-            value={page}
-            totalResults={patchTasks?.count}
-          />
-          <PageSizeSelector
-            data-cy="tasks-table-page-size-selector"
-            value={limit}
-            sendAnalyticsEvent={() => sendEvent({ name: "Change Page Size" })}
-          />
-        </TableControlInnerRow>
-      </TableControlOuterRow>
+      <TabTableControl
+        filteredCount={patchTasks?.count}
+        taskCount={taskCount}
+        limit={limit}
+        page={page}
+        defaultSortMethod={defaultSortMethod}
+      />
       {!data ? (
         <Skeleton active title={false} paragraph={{ rows: 8 }} />
       ) : (
@@ -147,13 +104,3 @@ const getQueryVariables = (
     limit: getLimitFromSearch(search),
   };
 };
-
-const FlexContainer = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-// @ts-expect-error
-const PaddedButton = styled(Button)`
-  margin-left: ${size.m};
-`;
